PM(misc): Push For Check
Uploading Current Code

The issue is that the PDF files downloads, but when trying to
open the download file, it states that the file is damaged. It
has to be something with how the file is being saved, because I
can open other PDFs on my Mac. Moreover, I've tried using a virtual
link being added to the DOM and being programatically clicked and
using the 'file-saver' library. Both methods do the download but
render the same issue.

I believe it's an issue with how Gatsby is storing the URL to the
PDF static asset.