UI(page): Edit Articles And Homepage Layout
Adjusting Grid On Each Page

On Homepage, the 9 most recent articles are fetched and
displayed as a 3x3 grid of ArticleCards. On /articles page
all blog articles are currently fetched and displayed in
a 2x2 grid.

Each blog template page should hold 20 articles. Pagination
should be added to fetch another 20 articles. Moreover, need
to find a way to build a Tag-And-Search system so users can
quickly look through articles that interest them.